restructure deepcrf. add deepcrf-policy in visualize
