+++ .gitignore
@@ -1,5 +1,6 @@
 # ignore tensoboard logs
 deepcrf/runs/
+runs/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
