+++ deepcrf/pokerenv_crf.py
@@ -2,24 +2,9 @@ import gym
 import numpy as np
 from gym import spaces
 from treys import Card, Deck, Evaluator
-from enum import Enum
 
+from enums import Action, Stage
 from obs_processor import ObsProcessor
-
-
-class Action(Enum):
-    FOLD = 0
-    CHECK_CALL = 1
-    RAISE = 2
-    ALL_IN = 3
-
-
-class Stage(Enum):
-    PREFLOP = 0
-    FLOP = 1
-    TURN = 2
-    RIVER = 3
-    END = 4
 
 
 def _convert_list_of_cards_to_str(cards):
