+++ deepcrf/train_deepcrf.py
@@ -5,12 +5,12 @@ import numpy as np
 from tqdm import tqdm
 import torch.multiprocessing as mp
 from torch.utils.tensorboard import SummaryWriter
+
+from model import BaseModel
+from player_wrapper import PolicyPlayerWrapper
 from pokerenv_crf import Action, HeadsUpPoker, ObsProcessor
 
-SUITS = 4
-RANKS = 13
 NUM_WORKERS = 16
-EMBEDDING_DIM = 128
 
 
 class AlwaysCallPlayer:
@@ -39,23 +39,6 @@ class ValuePlayerWrapper:
             obs = _batch_obses([obs])
             values = self.player(obs)
             action = torch.multinomial(regret_matching(values), 1).item()
-            return action
-
-
-class PolicyPlayerWrapper:
-    def __init__(self, player):
-        self.player = player
-
-    def __call__(self, obs):
-        with torch.no_grad():
-            obs = _batch_obses([obs])
-            action_distribution = self.player(obs)
-            action_distribution = torch.clamp(action_distribution, min=0)
-            total_action_distribution = torch.sum(action_distribution)
-            if total_action_distribution <= 0:
-                return np.random.choice(range(len(action_distribution)))
-            action_distribution /= total_action_distribution
-            action = torch.multinomial(action_distribution, 1).item()
             return action
 
 
@@ -121,65 +104,6 @@ class EvalPolicyPlayer:
             self.logger.add_scalar(
                 f"policy/{opponent_name}/mean_reward", np.mean(rewards), 0
             )
-
-
-class BaseModel(torch.nn.Module):
-    def __init__(self):
-        super(BaseModel, self).__init__()
-        self.num_actions = len(Action)
-        self.rank_embedding = torch.nn.Embedding(RANKS + 1, EMBEDDING_DIM)
-        self.suit_embedding = torch.nn.Embedding(SUITS + 1, EMBEDDING_DIM)
-        self.card_embedding = torch.nn.Embedding(RANKS * SUITS + 1, EMBEDDING_DIM)
-        self.stage_embedding = torch.nn.Embedding(4, EMBEDDING_DIM)
-        self.first_to_act_embedding = torch.nn.Embedding(2, EMBEDDING_DIM)
-        self.act = torch.nn.ReLU()
-        self.fc1 = torch.nn.Linear((15 + 6 + 1 + 1) * EMBEDDING_DIM + 8, 256)
-        self.fc2 = torch.nn.Linear(256, 128)
-        self.fc3 = torch.nn.Linear(128, self.num_actions)
-        self.fc3.weight.data.fill_(0)
-        self.fc3.bias.data.fill_(0)
-
-    def forward(self, x):
-        stage = x["stage"]
-        board_and_hand = x["board_and_hand"]
-        first_to_act_next_stage = x["first_to_act_next_stage"].long()
-
-        # B, 21
-        batch_size = board_and_hand.size(0)
-        board_and_hand = board_and_hand.view(batch_size, 7, 3)
-
-        ranks = board_and_hand[:, :, 0].long()
-        suits = board_and_hand[:, :, 1].long()
-        card_indices = board_and_hand[:, :, 2].long()
-
-        ranks_emb = self.rank_embedding(ranks)
-        suits_emb = self.suit_embedding(suits)
-        card_indices_emb = self.card_embedding(card_indices)
-
-        board_and_hand_emb = torch.cat([ranks_emb, suits_emb, card_indices_emb], dim=2)
-        board_and_hand_emb = board_and_hand_emb.view(batch_size, -1)
-
-        stage_emb = self.stage_embedding(stage)
-        first_to_act_next_stage_emb = self.first_to_act_embedding(
-            first_to_act_next_stage
-        )
-
-        all_emb = torch.cat(
-            [
-                board_and_hand_emb,
-                stage_emb,
-                first_to_act_next_stage_emb,
-                x["bets_and_stacks"],
-            ],
-            dim=1,
-        )
-
-        x = self.fc1(all_emb)
-        x = self.act(x)
-        x = self.fc2(x)
-        x = self.act(x)
-        x = self.fc3(x)
-        return x
 
 
 def _batch_obses(obses):
